requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2>=2.8.6,<2.8.7
flake8==3.9.2
coreapi==2.3.3
drf-yasg==1.20.0

